% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIS.gaussian.R
\name{SIS.gaussian}
\alias{SIS.gaussian}
\title{Sure Independence Screening}
\usage{
SIS.gaussian(X, Y, pred, scale = F)
}
\arguments{
\item{pred}{number of relevant variable to select, and \code{pred} is smaller than n. A conservative \code{pred} is \code{(n-1)} and also can try \code{nlog(n)}.}

\item{scale}{If scale=TRUE, X will be scaled.}

\item{x}{Ultra high dimensional dependent variables matrix.}

\item{y}{Continuous response vector if \code{family} is \code{"gaussian"}.}
}
\value{
A smaller dimensional set of regressors for further modeling, \code{AdaElasticNet} model object,
}
\description{
Sure Independence Screening
}
\examples{

x=matrix(rnorm(100*20),100,200)
y=rnorm(100)
Xsis=SIS.gaussian(x,y,pred = 20)
}
\references{
Jianqing Fan and Jinchi Lv (2008)
Sure Independence Screening for Ultrahigh Dimensional Feature Space (with discussion).
\emph{Journal of Royal Statistical Society} B, 70, 849-911.
}
\author{
Ziqi Zang
}
