% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.glmnet.R
\name{tune.glmnet}
\alias{tune.glmnet}
\title{Tune Hyperparameters for Adaptive Elastic Net}
\usage{
tune.glmnet(x, y, family, alphas, tune, nfolds, rule, ebic.gamma,
  lower.limits, upper.limits, seed, parallel, ...)
}
\arguments{
\item{x}{Data matrix.}

\item{y}{Response vector if \code{family} is \code{"gaussian"}.}

\item{family}{Model family, only \code{"gaussian"}.}

\item{alphas}{Vector of candidate \code{alpha}s to use in
\code{\link[glmnet]{cv.glmnet}}.}

\item{tune}{Parameter tuning method for each estimation step.
Possible options are \code{"cv"}, \code{"ebic"}, \code{"bic"},
and \code{"aic"}. Default is \code{"cv"}.}

\item{nfolds}{Fold numbers of cross-validation when \code{tune = "cv"}.}

\item{rule}{Lambda selection criterion when \code{tune = "cv"},
can be \code{"lambda.min"} or \code{"lambda.1se"}.
See \code{\link[glmnet]{cv.glmnet}} for details.}

\item{ebic.gamma}{Parameter for Extended BIC penalizing
size of the model space when \code{tune = "ebic"},
default is \code{1}. For details, see Chen and Chen (2008).}

\item{lower.limits}{Lower limits for coefficients.
Default is \code{-Inf}. For details, see \code{\link[glmnet]{glmnet}}.}

\item{upper.limits}{Upper limits for coefficients.
Default is \code{Inf}. For details, see \code{\link[glmnet]{glmnet}}.}

\item{init}{Type of the penalty used in the initial
estimation step. In the paper, they use \code{"enet"}.}
}
\value{
List of model coefficients, \code{glmnet} model object,
and the optimal parameter set.
}
\description{
Tune Hyperparameters for Adaptive Elastic Net
}
\examples{

x=matrix(rnorm(100*20),100,20)
y=rnorm(100)
fit=AdaElasticNet(x,y,alphas = seq(0.2, 0.8, 0.2), seed = 10)
print(fit)
plot(fit)
}
\references{
Zou, Hui, and Hao Helen Zhang. (2009).
On the adaptive elastic-net with a diverging number of parameters.
\emph{The Annals of Statistics} 37(4), 1733--1751.
}
\author{
Ziqi Zang
}
